Refactor into cleaner package structure
Moved classes into packages
