+++ app/src/main/AndroidManifest.xml
@@ -9,7 +9,7 @@
         android:icon="@drawable/ic_launcher"
         android:label="@string/app_name"
         android:theme="@style/Theme.DesignDemo" >
-        <activity android:name="com.designdemo.uaha.MainActivity" >
+        <activity android:name="com.designdemo.uaha.ui.MainActivity" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
@@ -18,14 +18,14 @@
             </intent-filter>
         </activity>
         <activity
-            android:name="com.designdemo.uaha.DetailActivity"
-            android:parentActivityName="com.designdemo.uaha.MainActivity" >
+            android:name="com.designdemo.uaha.ui.DetailActivity"
+            android:parentActivityName="com.designdemo.uaha.ui.MainActivity" >
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="com.designdemo.uaha.MainActivity" />
+                android:value="com.designdemo.uaha.ui.MainActivity" />
         </activity>
         <activity
-            android:name="com.designdemo.uaha.UserActivity"
+            android:name="com.designdemo.uaha.ui.UserActivity"
             android:label="@string/title_activity_user" >
         </activity>
     </application>
