+++ app/src/main/java/com/designdemo/uaha/ui/ProductListFragment.java
@@ -1,4 +1,4 @@
-package com.designdemo.uaha;
+package com.designdemo.uaha.ui;
 
 import android.app.Activity;
 import android.content.Context;
@@ -6,6 +6,10 @@ import android.content.Intent;
 import android.os.Build;
 import android.os.Bundle;
 import androidx.annotation.Nullable;
+
+import com.designdemo.uaha.ui.common.GridDividerDecoration;
+import com.designdemo.uaha.data.VersionData;
+import com.designdemo.uaha.util.PrefsUtil;
 import com.google.android.material.bottomsheet.BottomSheetBehavior;
 import androidx.core.app.ActivityOptionsCompat;
 import androidx.fragment.app.Fragment;
